# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "chat-API",
    "description": "A little whatsapp api clone...",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "login user",
        "operationId": "loginuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginuserrequest"
                  },
                  {
                    "example": {
                      "email": "J-NoCopyright@outlook.com",
                      "password": "Academlo123@"
                    }
                  }
                ]
              },
              "example": {
                "email": "J-NoCopyright@outlook.com",
                "password": "Academlo123@"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/messages/conversation/{id}": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "create new messages",
        "operationId": "createnewmessages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 4
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createnewmessagesrequest"
                  },
                  {
                    "example": {
                      "content": "queti",
                      "senderId": 1
                    }
                  }
                ]
              },
              "example": {
                "content": "queti",
                "senderId": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "get conversation messages",
        "operationId": "getconversationmessages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 4
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "loginuserrequest": {
        "title": "loginuserrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "J-NoCopyright@outlook.com",
          "password": "Academlo123@"
        }
      },
...
      "createnewmessagesrequest": {
        "title": "createnewmessagesrequest",
        "required": [
          "content",
          "senderId"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "senderId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "content": "queti",
          "senderId": 1
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}